backend/requirements.txt
django<4
djangorestframework==3.14.0
pycodestyle==2.10.0